+++ selfdrive/ui/sunnypilot/qt/onroad/hud.cc
@@ -108,6 +108,7 @@ void HudRendererSP::updateState(const UIState &s) {
 
   if (sm.updated("deviceState")) {
     memoryUsagePercent = sm["deviceState"].getDeviceState().getMemoryUsagePercent();
+    freeSpacePercent = sm["deviceState"].getDeviceState().getFreeSpacePercent();
   }
 
   if (sm.updated(gps_source)) {
@@ -478,6 +479,9 @@ void HudRendererSP::drawBottomDevUI(QPainter &p, int x, int y) {
 
   UiElement memEl = DeveloperUi::getMemoryUsagePercent(memoryUsagePercent); // 内存占用
   rw += drawBottomDevUIElement(p, rw, y, memEl.value, memEl.label, memEl.units, memEl.color);
+
+  UiElement storageEl = DeveloperUi::getFreeSpacePercent(freeSpacePercent); // 存储剩余
+  rw += drawBottomDevUIElement(p, rw, y, storageEl.value, storageEl.label, storageEl.units, storageEl.color);
 
   UiElement maxCpuTempElement = DeveloperUi::getMaxCpuTemp(maxCpuTemp); // cpu温度
   rw += drawBottomDevUIElement(p, rw, y, maxCpuTempElement.value, maxCpuTempElement.label, maxCpuTempElement.units, maxCpuTempElement.color);