+++ selfdrive/ui/sunnypilot/qt/onroad/hud.h
@@ -51,6 +51,7 @@ private:
   float roll;
   int memoryUsagePercent;
   float cpuUsagePercent;
+  int freeSpacePercent;
   int devUiInfo;
   float gpsAccuracy;
   float altitude;